+++ src/components/common/CollapseCard.vue
@@ -1,5 +1,5 @@
 <template>
-  <div :class="`collapse ${showCollapse ? 'collapse-open' : 'collapse-close'}`">
+  <div :class="`group collapse ${showCollapse ? 'collapse-open' : 'collapse-close'}`">
     <div
       class="collapse-title cursor-pointer pr-4"
       @click="showCollapse = !showCollapse"