+++ src/components/settings/general/ZashboardSettings.vue
@@ -38,19 +38,7 @@
       >
         {{ $t('upgradeDashboard') }}
       </button>
-      <button
-        class="btn btn-sm"
-        @click="handlerClickResetSettings"
-      >
-        {{ $t('resetSettings') }}
-      </button>
-      <button
-        class="btn btn-sm"
-        @click="exportSettings"
-      >
-        {{ $t('exportSettings') }}
-      </button>
-      <ImportSettings />
+      <DashboardSettings />
     </div>
 
     <StyleSettings />
@@ -63,11 +51,9 @@ import { upgradeUIAPI, zashboardVersion } from '@/api'
 import { useIsSettingVisible, useSettings } from '@/composables/settings'
 import { GENERAL_ITEM_KEYS } from '@/config/settingsItems'
 import { handlerUpgradeSuccess } from '@/helper'
-import { exportSettings, resetSettings } from '@/helper/utils'
 import { twMerge } from 'tailwind-merge'
 import { ref } from 'vue'
-import { useI18n } from 'vue-i18n'
-import ImportSettings from '../../common/ImportSettings.vue'
+import DashboardSettings from '../../common/DashboardSettings.vue'
 import GeneralSettings from './GeneralSettings.vue'
 import StyleSettings from './StyleSettings.vue'
 
@@ -77,14 +63,8 @@ const isVisibleActions = useIsSettingVisible(k.actions)
 const commitId = __COMMIT_ID__
 
 const { isUIUpdateAvailable } = useSettings()
-const { t } = useI18n()
 
 const isUIUpgrading = ref(false)
-
-const handlerClickResetSettings = () => {
-  if (!window.confirm(t('resetSettingsConfirm'))) return
-  resetSettings()
-}
 
 const handlerClickUpgradeUI = async () => {
   if (isUIUpgrading.value) return